# perf: improve solve_dependencies increasing requests/s by 18%