@yezz123 2022-10-14
> Tests are passing now, but the docs generation was canceled and I have no way to retry the workflow without pushing unnecessary commits.

Lately some issues are related to the packages and containers we use Github actions that's why, but overall no problem if the PR doesn't have effect in docs 🚀