@Nikakto 2022-11-06
Does anybody know reasons why fastapi resolve dependencies one by one instead like asyncio.gather() ?

Code example:
```
@pytest.mark.asyncio
async def test_fastapi_is_slowapi():
    async def dep0():
        await asyncio.sleep(5)
        return 0

    async def dep1():
        await asyncio.sleep(5)
        return 0

    @app.get("/testslow")
    async def app_testslow(d0=Depends(dep0), d1=Depends(dep1)) -> int:
        return 10

    response = await client.get("/testslow")

    assert response.elapsed.seconds >= 10  # True
```