@sk- 2024-02-10
I'm closing this PR because the fastapi team does not seem to be interested in this at all.

I took into consideration all requested changes, simplifying the PR, and commented on this issue and similar but no one from the core team has even commented on whether they are willing to merge it. At some point they marked it as investigating, but nothing was done. They didn't even requested further performance testing or benchmark.

Recently, they marked it as P4, without even mentioning why. This is surprising because it's not even a complicated PR.

So, in light of the team's (lack of) response, I'm closing this.